x86/mm: dedup the various copies of the shadow lock functions
Define the lock and unlock functions once, and list all the locks in one
place so (a) it's obvious what the locking discipline is and (b) none of
the locks are visible to non-mm code.  Automatically enforce that these
locks never get taken in the wrong order.

Signed-off-by: Tim Deegan <Tim.Deegan@citrix.com>
(cherry picked from commit 301493fb027648db6808b66d1ccf849f524b8422)
Signed-off-by: Bastian Blank <bastian.blank@credativ.de>